+++ src/components/multi-autocompleter/multi-autocompleter.test.jsx
@@ -302,4 +302,31 @@ describe('<MultiAutocompleter>', () => {
       expect(screen.getByRole('combobox', { name: requiredProps.label })).toHaveAccessibleDescription('');
     });
   });
+
+  describe('onFocus API', () => {
+    it('triggers callback when tooltip is focused', () => {
+      const onFocus = jest.fn(event => event.persist());
+      render(<MultiAutocompleter {...requiredProps} onFocus={onFocus} />);
+
+      userEvent.click(screen.queryByRole('combobox'));
+      expect(onFocus).toHaveBeenCalledWith(expect.objectContaining({
+        type: 'focus',
+      }));
+    });
+  });
+
+  describe('onBlur API', () => {
+    it('triggers callback when tooltip is focused', () => {
+      const onBlur = jest.fn(event => event.persist());
+      render(<MultiAutocompleter {...requiredProps} onBlur={onBlur} />);
+
+      userEvent.click(screen.queryByRole('combobox'));
+      userEvent.tab(); // Clear Button `X`
+      userEvent.tab(); // Expend Button `V`
+      userEvent.tab(); // Document Body
+      expect(onBlur).toHaveBeenCalledWith(expect.objectContaining({
+        type: 'blur',
+      }));
+    });
+  });
 });